Add overloaded_service_name and common_tags parameters to datadog plugin.
These two parameters are optionnal and do not change the current behavior by default.

'overloaded_service_name' allows to force the string used in place of the service name when building the metrics name. This is useful when several kong services actually implement the same "conceptual" service, for instance when several upstreams provide the same API. For cases like this, datadog encourages the usage of a common metric with tags to filter the subsets when required.

'common_tags' allows to specify a collection of tags at service level that will be used by all metrics. The common tags are added to the metrics-level tags. This is actually a sugar parameter to avoid redefining all the default metrics just to apply the same tag to all of them. This helps keep the configuration logic simple and the configuration files (when using DB-less mode) more readable.